Limit access for tasks to each creating user

+++ db/models/Task.js
@@ -27,6 +27,8 @@ exports.addTasks = (tasks, userId) => {
     .then(res => res.rows.map(row => row.id));
 };
 
-exports.toggleTask = taskId => pool.query('UPDATE tasks SET completed = NOT completed WHERE id = $1', [taskId]);
+exports.toggleTask = (taskId, userId) => pool.query('UPDATE tasks SET completed = NOT completed WHERE id = $1 AND users_id = $2', [taskId, userId])
+  .then(res => res.rowCount > 0);
 
-exports.deleteTask = taskId => pool.query('DELETE FROM tasks WHERE id = $1', [taskId]);
+exports.deleteTask = (taskId, userId) => pool.query('DELETE FROM tasks WHERE id = $1 AND users_id = $2', [taskId, userId])
+  .then(res => res.rowCount > 0);
